Rename(HomeController, ApiHomeController, ApiAdminController)-> 프로젝트 표기법에 맞게 이름 수정후 HomeController에 있는 api기능 ApiHomeController로 뺴서 다른패키지와같이 통일화함
