+++ src/main/java/kr/kro/bbanggil/newsletter/service/NewsletterService.java
@@ -7,4 +7,5 @@ public interface NewsletterService {
 	
 	
 	NewsletterResponseDto getLatestNewsletter();
+	 public String calculateNextNewsletterSchedule();
 }
