+++ src/main/java/kr/kro/bbanggil/newsletter/service/NewsletterServiceImpl.java
@@ -1,5 +1,11 @@
 package kr.kro.bbanggil.newsletter.service;
 
+import java.time.DayOfWeek;
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+import java.time.format.DateTimeFormatter;
+
 import org.springframework.stereotype.Service;
 
 import kr.kro.bbanggil.admin.dto.response.NewsletterResponseDto;
@@ -20,4 +26,15 @@ public class NewsletterServiceImpl implements NewsletterService{
 		
 		return   newsletterMapper.selectLatestNewsletter();
 	}
+	
+	 @Override
+	    public String calculateNextNewsletterSchedule() {
+	        LocalDate today = LocalDate.now();
+	        int daysUntilNextMonday = (DayOfWeek.MONDAY.getValue() - today.getDayOfWeek().getValue() + 7) % 7;
+	        if (daysUntilNextMonday == 0) daysUntilNextMonday = 7;
+
+	        LocalDate nextMonday = today.plusDays(daysUntilNextMonday);
+	        LocalDateTime nextNewsletterTime = LocalDateTime.of(nextMonday, LocalTime.of(9, 0));
+	        return nextNewsletterTime.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"));
+	    }
 }
